Store Support/General Setup/BigQueryProject/requirements.txt
# Core BigQuery Integration
google-cloud-bigquery>=3.0.0
google-cloud-functions-framework>=3.0.0
google-cloud-logging>=3.0.0
google-auth>=2.0.0

# Data Processing & Analysis
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0

# Visualization & Reporting
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# File Format Support
openpyxl>=3.0.0
pyarrow>=8.0.0

# Optional: Advanced Analytics
sqlalchemy>=1.4.0
db-dtypes>=1.0.0
